// src/common/types/redisSeasonLevel.ts
type RedisSeasonLevel = {
    level: number;
    minPoints: number;
    maxPoints: number;
	display: {
		description: string;
		textColor: string;
		accentColor: string;
		textAccentColor: string;
		cardBackgroundImage: string;
		backgroundColor: string;
		backgroundImage: string;
	}
}

export default RedisSeasonLevel;